Initial Commit - Portfolio!
This has been on the radar for years now.

I think the biggest obstacle was more of a personal one to overcome. The confidence in myself to be able to say the things I'm good at, as well as being open about myself in that way. It was about the worry of being boastful or self-centered. I also feel I have a lot of things in my 'personal catalog' to show for now, and I feel it's worth talking about. I am proud of myself and who I am now, and I want to be able to speak that. It's not about getting others' approval, it's about being there for yourself and putting yourself out there because you know you can be a viable asset to someone else who you might be able to work with or help.