Hello World 🌎
Realized the main app didn't really have anything to start with, some of the styles were odd too, mainly because the skeleton of this I borrowed from my Dovetail project, just with updated dependency versions and the Service Worker removed. I plan to re-add that later on maybe, but we'll see. I don't know if this really needs to be a PWA. I like the idea of it being a classic plain kind of website, just a blog or something like that. It is meant to be a showcase for sure though.