GH Action Updates
Still a learning curve for me in how to use these. I am not a fan of this not being documented too well, maybe I am just missing something. I like that for example npm shows you when something is out of date, this just plain doesn't work when you try to push.