Full TS Paths
I was not in favor of this for the longest time because it wasn't very "browser-like", since paths didn't resolve to real files. In this context though, I am now okay with things being "magic", since now I know for certain that I want to use bundling here. For things like npm packages though, you don't want to use magic paths, at least I feel, because those will be used as raw files for the consumer, it's up to their bundler to transform your code from plain ESM to that of something else. I feel it makes more sense to distribute package code in a vanilla manner, with only using custom module paths where absolutely necessary. Bundling and or Node.js shouldn't be expected for the end-all-be-all by which the package will be used. It might just be plain vanilla JS, so one should account for that (especially now that we have things like `await import()` for example).