Central Components (Skill Level Reflections)
Looking into more info about SolidStart, Prettier, ESLint, and typescript-eslint. I very much value the things I've learned over in the Rust world when using the tooling for that. It seems JS is still a bit fragmented in that regard though, because all of the tools are implemented on their own, and hooked together to be used as one. Say for example, installing ESLint, Prettier, and TypeScript on their own. In Rust, you just have `rustfmt` and `rustc`. Not to say JS should be like this too, but I think what I am realizing is that JS is still not completely in a deterministic state, because there are still a few instances of bikeshedding for certain opinions and community disagreements. For example, I still like having no-space closed JSX tags, but it seems like the convention is to still have the space, not because of the history behind the syntax being one way or another, but instead because it was what people are used to. JS is stuck in the sweet spot of wanting to be deterministic like Rust, but is very much not based in that kind of thing, so I think it can't be. Rust is not JS. Nor is npm Cargo. JS is awesome for what it is, as is Rust. I am going to continue using JS like JS, rather than worrying about trying to make it deterministic like Rust. The web is a great place, I'm tired of worrying about what it is or isn't, I'm going to just go back to making the things I want to make, not worrying about what people think or what I think. It doesn't matter, it's a tool to create projects that you see. That's the important part.

Gotta say, having Vite hot reload tied in with VSCode auto-save (300ms), it is very much nearly exactly like what I wanted to have with STE years back. I am now at that point of ease again, with even more potential under my belt now. It's no longer limited by editing JS in plain text, now I have types, and the frameworks I am using to build this allow me to move much faster and efficiently, allowing for much better code quality and practices at the same time. Dependency injection, functional paradigms, clean module graphs (no recursion), and coupling that with effect-based UI driven by signal-based state hooks, I am unstoppable! WATCH OUT!! 🏎️ 🏁

Okay, so I think all of my issues from a few years ago just resolve down to skill issue