Patches Update
Bringing this to the next build before I move on to adding more things to the app specifically.

Felt kind of mind-numb after using the energy to write all of the stuff in the previous messages, as well as reading the docs for things. Learning how to recognize when I need to take a break to refresh my mind and motives again. I tend to get fixated on just keeping working, even when fatigued. I thought stopping for breaks was just making the process longer. Now I see that not stopping just makes my performance dip. This is present in more than one aspect for me. I used to see it make more sense to continue through the night to finish things, even if that meant not sleeping in a regular schedule. Now I can see how irregular timings can change my performance in terms of things like exercise, sleep deprevation, hunger, and mental stamina. I need sleep, food, water, and emotional honesty. With those things in check, I can then sit down and concentrate on making the things I want to make, and being a good person. I tend to get more self-critical and down on low sleep too.

I could keep going with these findings. That was a tangent. The short answer for this commit: Took a 10 minute snack break because I realized my mind was starting to lock up, came back to start the next iteration of this project.