Self-Showcase!
Working on learning how to present myself to the world! I realized while writing this, that I still have high tendencies to need to explain myself before I am able to just show the work. I do this in real-life too, prefacing all of my opinions and work with reasons for why it might be right-wrong, and why it might not be something you like-dislike, as well as something you might expect-not expect. I am now starting to see that I don't need to do this. It is distracting once again from that initial impact. I am afraid of how I am seen. I have a voice. I am afraid to use it.

Signpost to the Stars - Feed the Forest (Daisy Rickman)

I am not ashamed of this! Me getting to this point in the first place is progress. All I am looking for is progress. The outcomes come as a result of progress. You can't create outcomes, you can only shape progress in the now.