Public Data Folder
Okay now the data itself is in the public folder. I want this to be more accessible in terms of the build itself. I thought I initially wanted to bundle it with the app itself, but it might actually make sense just to have this as plain JSON. I'll test whether performance for this makes sense.