Root Data Path
This should be an absolute path, rather than a relative one, Vite mentioned a recommendation for this in the terminal. I've run into this before, I think I usually start with relative paths because then you are certain in VSCode that the path is correct when you initially try to import it. That's one thing about Vite typings for TS, it doesn't resolve to real folder typings, it is just a type annotation. So it works great for files you know are there, but otherwise they are like an assertion, it could be incorrect if the file isn't there. You would only discover it in runtime, which is less than ideal. Same with CSS imports.