Spring Release Update
New releases I've been publishing and working on recently. Thought I'd bring them to this project too.

I am really looking forward to making a graph that shows how long each album of mine "takes to make", as well as when it finally is "finished", compared to when it is "released". These are three dates essentially. The only metric I don't have is when the concept for it started. I don't know when that starts, let alone how to keep track of that. These dates are also only able to be viewed after they happen. I don't plan for specific ones.

The only thing I do know is that I seem to be going on a personal trend (not like online trend, but a pattern I am doing) where I seem to be releasing more things as they are "finished", rather than waiting to feel for when they should be released. Maybe that's more of a reflection of my ability to stay a little less close to the work, in terms to how I feel about it. The perfectionism is much lower now, and instead I am curious about learning how to do things beforehand, then just seeing what comes of that afterwards. So it's like it's front-loading the end result.

To be fair though, the things I am releasing quickly now, they have formed the way they have because of the complex releases I am still figuring out what they mean to be internally. Because those had their ways of explaining or expressing things, that's what allowed these to flow free-er. So maybe it's more that the result of that complex work, that these releases are ready to come out.

Maybe there isn't really a pattern of timing though, instead more just about the abstractness of it, or what each project stems from. I have a bunch of songs I know are "done", but the artwork or name might not have been apparent when making it. Same with the sequencing, if it needed it. Or even the confidence I had while thinking of it as a final product. These things seem to be why the current things are releasing fast. For these, I seem to know what they are, and what criteria they are allowed to fill into. These ones have to be a certain way, rather than the other way around.

The three new tracks for example, those will be a single album eventually. I am still writing for it right now, I can tell. How I arrange it will come with time. The songs themselves I know are already done though, and I now seem to be capable with releasing them as pre-release singles. I hadn't really done this before. It feels great.

Trashmouth needed a bit more work to figure out what it sounded like though, and it really way composed of two one-week projects. I realized that they were really a single time period, and they have a lot of overlap. They seem like two halves to the same complete story. The sequencing matches how they were composed as well. That release seemed to work in that way.

I think what seems to take time is that realization of what it is, as well as time to step back and see it for what it was, rather than what it is. Seeing it from the past seems to make it more objective as a project, rather than an expression of you right now.

The new singles are the opposite though, they are really very much meant to show me, right now. That is the unique goal of that specific project. I didn't decide that intentionally, it just seems like that is the answer as to what is next.

There are a bunch of projects and concepts I "want" to do. But when I do those irrespective of what I can make right in the current moment, this is where it gets messy. I am now looking into not doing that. I am hoping to try making only what I am capable of making right now. This has been the common theme for my projects here and there, but sometimes it doesn't get acknowledged, and I hadn't know this outright as something I had to think about.